+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 # Lamp Tools
 
+## v0.1.1
+
+### Added
+
+- The following arguments:
+  - `-h, --help`
+  - `--setup-dev-env`
+  - `--setup-apache2`
+  - `--setup-php`
+  - `--setup-mysql`
+
+### Changed
+
+- Merged in the various tools used in the script `setup-dev-env.sh` into this script. #1
+
+### Issues Closed
+
+- #1
+
+---
+
 ## v0.1.0
 
 ### Added